[CMP-1298] Add helm linting (#5)
* 🎉 feat(.github/workflows/helm-lint.yml): add GitHub action for Helm linting on pull requests
A new GitHub action has been added to perform Helm linting whenever a pull request is made. This ensures that any changes made to the Helm charts are valid and follow the correct syntax before they are merged into the main branch. This will help maintain the quality of the code and prevent potential issues that could arise from invalid Helm charts.

* 🔧 chore(github-actions): add new ActionLint workflow and update Helm Lint workflow
The new ActionLint workflow is added to ensure that all GitHub Actions workflows are correctly formatted and follow best practices. This will help to catch any potential issues before they are merged into the main branch. The Helm Lint workflow is updated to use the latest version of the checkout action (v4) and to include the GitHub token in the setup step. This will ensure that the workflow has the necessary permissions to perform its tasks.