🎨 refactor(release-helm-chart): extract release helm chart steps into a reusable GitHub action The steps to release a Helm chart were extracted into a separate GitHub action to promote reusability and maintainability. This change allows the release process to be used in multiple workflows without duplicating code. It also makes the workflow file cleaner and easier to read.
📝 docs: add Apache License 2.0
The Apache License 2.0 has been added to the project to provide clear terms and conditions for use, reproduction, and distribution of the project's code. This license is widely used due to its flexibility and legal clarity, and it allows for the code to be used in free and open source software projects as well as proprietary ones.