[CMP-1299] Fix github token access
🔧 refactor(action.yml): use create_branch step output for PR head reference to improve code readability
The github_token is now passed to each step of the GitHub Actions workflow to authenticate the GitHub Actions bot. This ensures that the bot has the necessary permissions to perform the actions. The head reference for the pull request is now using the output from the create_branch step, which improves code readability by making it clear where the branch name is coming from.