[CMP-1299] Combine push and pr create
The pull request creation step was removed from the GitHub action for releasing the Helm chart. This change was made to streamline the release process. Now, the action directly commits the changes to the main branch, eliminating the need for a pull request. This simplifies the workflow and speeds up the release process.