[CMP-1299 ] Specify head for PR
The branch id was added to the "Create New Branch for Changes" step to capture the branch name. This branch name is then used in the PR creation command as the head parameter. This change ensures that the correct branch is used when creating the PR, preventing potential errors or conflicts.