[CMP-1299] Create the cr-release-packages directory if it does not exist
✨ feat(release-helm-chart/action.yml): add .cr-release-packages directory creation step
The skip_existing and skip_upload options were removed from the Publish Helm chart action to ensure that all packages, even those that already exist, are processed and uploaded. This change ensures that no package is accidentally skipped during the release process. Additionally, a step to create the .cr-release-packages directory was added to ensure that the directory exists before the release process begins, preventing potential errors.