[CPM-1299]: replace helm/chart-releaser-action with manual chart releaser installation and execution (#23)
The helm/chart-releaser-action was replaced with a manual installation and execution of the chart releaser. This change provides more control over the release process and allows for additional steps such as packaging the Helm chart and creating a GitHub release. This change also includes the addition of a new binary file for the Helm chart. The manual process ensures that the Helm chart is correctly packaged and released on GitHub, providing a more reliable deployment process.