[CMP-1299]: remove redundant 'cr --version' command from Install Chart Releaser step (#25)
The 'cr --version' command was removed from the Install Chart Releaser step in the GitHub action for releasing Helm charts. This command was redundant as it was only used for debugging purposes and does not affect the functionality of the action. Removing it simplifies the action and reduces unnecessary output in the action logs.