[CMP-1299]: update cloudbolt-collector binary and index.yaml with new digest and timestamps (#40)
The cloudbolt-collector binary has been updated, which resulted in a new digest in the index.yaml. The timestamps for 'created' and 'generated' fields in index.yaml have also been updated to reflect the time of this change. This ensures that the documentation and the binary are in sync and up-to-date.