[CMP-1299] Fix helm chart signing (#41)
* 🔒 feat(release-helm-chart): add PGP signing to Helm chart release process
📦 fix(cloudbolt-collector): update binary files for cloudbolt-collector
The changes were made to enhance the security of the Helm chart release process by adding PGP signing. This ensures that the Helm charts are verified and trusted, reducing the risk of deploying tampered charts. The binary files for cloudbolt-collector were updated to reflect the latest changes in the application.

* 📝 docs(index.yaml): update created and generated timestamps and digest in cloudbolt-collector entry
The timestamps and digest in the cloudbolt-collector entry were updated to reflect the most recent changes. This ensures that the documentation is always up-to-date with the latest version of the application.