[CMP-1299] Use inputs not secrets (#45)
* 🔧 refactor(release-helm-chart/action.yml): redirect output results to GITHUB_STEP_SUMMARY
🔨 fix(release-helm-chart/action.yml): add shell specification for Install Helm step
The output results are now redirected to GITHUB_STEP_SUMMARY to provide a more organized and readable output. The shell specification for the Install Helm step was added to ensure the correct shell environment is used for the script execution.

* 🐛 fix(release-helm-chart/action.yml): correct syntax error in Import PGP Key step
The syntax error in the Import PGP Key step of the GitHub action was causing the action to fail. The 'bash: shell' and 'rum:' lines were corrected to 'shell: bash' and 'run:', respectively. This change ensures that the action runs as expected.

* 🔒 fix(release-helm-chart): use input for PGP private key instead of secret
The PGP private key is now passed as an input to the GitHub action instead of being stored as a secret. This change enhances the security of the action by not exposing the private key in the repository secrets, and it also provides more flexibility as the key can be changed per run without needing to modify the repository secrets.