[CMP-1299] fix keyring (#51)
* 🔧 refactor(release-helm-chart/action.yml): redirect output results to GITHUB_STEP_SUMMARY
🔨 fix(release-helm-chart/action.yml): add shell specification for Install Helm step
The output results are now redirected to GITHUB_STEP_SUMMARY to provide a more organized and readable output. The shell specification for the Install Helm step was added to ensure the correct shell environment is used for the script execution.

* 🐛 fix(release-helm-chart/action.yml): correct syntax error in Import PGP Key step
The syntax error in the Import PGP Key step of the GitHub action was causing the action to fail. The 'bash: shell' and 'rum:' lines were corrected to 'shell: bash' and 'run:', respectively. This change ensures that the action runs as expected.

* 🔒 fix(release-helm-chart): use input for PGP private key instead of secret
The PGP private key is now passed as an input to the GitHub action instead of being stored as a secret. This change enhances the security of the action by not exposing the private key in the repository secrets, and it also provides more flexibility as the key can be changed per run without needing to modify the repository secrets.

* 🔒 fix(action.yml): modify PGP key import command to support non-interactive mode
The PGP key import command has been updated to support non-interactive mode. This change is necessary to prevent the action from hanging when it requires user interaction. The `--batch` flag is used to ensure the command runs in non-interactive mode, and `--pinentry-mode loopback` allows GPG to read the passphrase from the command line or from a file. This makes the GitHub action more robust and reliable.

* 🔄 refactor(release-helm-chart): replace manual GPG key import with ghaction-import-gpg action
The manual import of the GPG key has been replaced with the ghaction-import-gpg action. This change was made to simplify the process and reduce the potential for errors. The ghaction-import-gpg action is more reliable and easier to maintain.

* 🔧 refactor(release-helm-chart/action.yml): split echo and helm package commands into separate lines for better readability
The echo and helm package commands were previously on the same line, which made the code harder to read. By splitting them into separate lines, the code is now more readable and easier to understand.

* 🔧 refactor(release-helm-chart): remove unnecessary backslashes in action.yml
The backslashes at the end of the lines were removed to improve readability and maintainability. They were unnecessary as the command arguments are already separated by newlines.

* 🔧 chore(release-helm-chart): add step to downgrade GPG to secring.pgp
The GPG version used in the GitHub action was causing compatibility issues. To resolve this, a step was added to downgrade GPG to secring.pgp. This ensures that the GPG version used is compatible with the rest of the workflow.

* 🔑 chore(release-helm-chart): add keyring path to helm chart release action
The keyring path is added to the helm chart release action to ensure that the correct keyring is used when signing the helm chart. This change is necessary to avoid any potential issues with keyring selection during the signing process.