[CMP-1299] Allow PR title to pass for Release PRs as well as Jira ticket PRs (#54)
* 🔧 refactor(jira-check.yml): rename job and split checks into separate steps for better readability and maintainability
✨ feat(jira-check.yml): add check for 'Release' prefix in PR title to enforce release naming convention
The job name has been changed from 'Ensure Jira Ticket' to 'Validate PR Title' to better reflect its purpose. The checks for Jira ticket and 'Release' prefix in PR title have been split into separate steps to improve readability and maintainability of the workflow. The addition of the 'Release' prefix check enforces a new naming convention for PRs related to releases, ensuring that they are easily identifiable.

* 🔥 remove(release-helm-chart.yml): Delete redundant GitHub workflow file
The release-helm-chart.yml GitHub workflow file was removed as it was no longer needed. This could be due to a change in the deployment process, or the tasks performed by this workflow have been moved to another workflow. This helps to keep the project clean and avoid confusion from having unnecessary files.

* 🐛 fix(jira-check.yml): modify JIRA_TICKET_REGEX to correctly validate JIRA ticket format
The JIRA ticket validation regex was updated to correctly validate the JIRA ticket format in the PR title. The previous regex was not correctly validating the format due to the caret (^) symbol at the start, which was causing it to only match the ticket format if it was at the beginning of the title. The updated regex will now correctly validate the ticket format anywhere in the title.